mqtt-client-py/requirements.txt
paho-mqtt==1.4.0
requests==2.22.0
pymongo==3.8.0
